fix(checker): keep bounded indexed recursion finite
