+++ crates/tsz-checker/src/types/type_checking/type_alias_checking.rs
@@ -790,6 +790,14 @@ impl<'a> CheckerState<'a> {
                 if self.type_arg_nodes_all_are_deferred_passthrough_for_depth_check(type_args) {
                     return false;
                 }
+                if type_args.nodes.iter().copied().all(|arg_idx| {
+                    self.type_node_is_deferred_passthrough_for_depth_check(arg_idx)
+                        || self.type_node_is_bounded_indexed_descent_for_depth_check(
+                            alias_sid, arg_idx,
+                        )
+                }) {
+                    return false;
+                }
                 return !self
                     .type_arg_nodes_contain_scoped_type_parameter_for_depth_check(type_args);
             }
@@ -826,6 +834,9 @@ impl<'a> CheckerState<'a> {
                             .type_arg_nodes_all_are_deferred_passthrough_for_depth_check(type_args)
                         && type_args.nodes.iter().copied().any(|arg_idx| {
                             !self.type_node_is_deferred_passthrough_for_depth_check(arg_idx)
+                                && !self.type_node_is_bounded_indexed_descent_for_depth_check(
+                                    alias_sid, arg_idx,
+                                )
                         })
                 });
         }
@@ -908,55 +919,6 @@ impl<'a> CheckerState<'a> {
             })
     }
 
-    fn type_arg_nodes_all_are_deferred_passthrough_for_depth_check(
-        &mut self,
-        type_args: &tsz_parser::parser::NodeList,
-    ) -> bool {
-        !type_args.nodes.is_empty()
-            && type_args
-                .nodes
-                .iter()
-                .copied()
-                .all(|node_idx| self.type_node_is_deferred_passthrough_for_depth_check(node_idx))
-    }
-
-    fn type_node_is_deferred_passthrough_for_depth_check(&mut self, node_idx: NodeIndex) -> bool {
-        let Some(node) = self.ctx.arena.get(node_idx) else {
-            return false;
-        };
-
-        if let Some(identifier) = self.ctx.arena.get_identifier(node)
-            && self
-                .ctx
-                .type_parameter_scope
-                .contains_key(&identifier.escaped_text)
-        {
-            return true;
-        }
-        if let Some(identifier) = self.ctx.arena.get_identifier(node)
-            && self
-                .identifier_references_enclosing_infer_binding(node_idx, &identifier.escaped_text)
-        {
-            return true;
-        }
-
-        if node.kind == syntax_kind_ext::TYPE_REFERENCE
-            && let Some(type_ref) = self.ctx.arena.get_type_ref(node)
-        {
-            if type_ref
-                .type_arguments
-                .as_ref()
-                .is_some_and(|type_args| !type_args.nodes.is_empty())
-            {
-                return false;
-            }
-
-            return self.type_name_is_deferred_passthrough_for_depth_check(type_ref.type_name);
-        }
-
-        false
-    }
-
     fn conditional_body_has_unresolved_computed_recursive_alias_ref(
         &mut self,
         node_idx: NodeIndex,
@@ -994,120 +956,6 @@ impl<'a> CheckerState<'a> {
                     child_idx, alias_sid,
                 )
             })
-    }
-
-    fn type_node_contains_unresolved_type_reference_for_depth_check(
-        &mut self,
-        node_idx: NodeIndex,
-    ) -> bool {
-        let Some(node) = self.ctx.arena.get(node_idx) else {
-            return false;
-        };
-
-        if node.kind == syntax_kind_ext::TYPE_REFERENCE
-            && let Some(type_ref) = self.ctx.arena.get_type_ref(node)
-            && self
-                .resolve_type_symbol_for_lowering(type_ref.type_name)
-                .is_none()
-            && self
-                .ctx
-                .arena
-                .kind_at(type_ref.type_name)
-                .is_some_and(|kind| kind == syntax_kind_ext::QUALIFIED_NAME)
-            && !self.type_name_is_deferred_passthrough_for_depth_check(type_ref.type_name)
-        {
-            return true;
-        }
-
-        self.ctx
-            .arena
-            .get_children(node_idx)
-            .into_iter()
-            .any(|child_idx| {
-                self.type_node_contains_unresolved_type_reference_for_depth_check(child_idx)
-            })
-    }
-
-    fn type_name_is_deferred_passthrough_for_depth_check(&mut self, name_idx: NodeIndex) -> bool {
-        let Some(name_node) = self.ctx.arena.get(name_idx) else {
-            return false;
-        };
-        let Some(identifier) = self.ctx.arena.get_identifier(name_node) else {
-            return false;
-        };
-
-        self.ctx
-            .type_parameter_scope
-            .contains_key(&identifier.escaped_text)
-            || self
-                .identifier_references_enclosing_infer_binding(name_idx, &identifier.escaped_text)
-    }
-
-    fn identifier_references_enclosing_infer_binding(
-        &self,
-        node_idx: NodeIndex,
-        name: &str,
-    ) -> bool {
-        let mut current = node_idx;
-        for _ in 0..50 {
-            let parent = self
-                .ctx
-                .arena
-                .get_extended(current)
-                .map_or(NodeIndex::NONE, |ext| ext.parent);
-            if parent.is_none() {
-                return false;
-            }
-            let Some(parent_node) = self.ctx.arena.get(parent) else {
-                return false;
-            };
-            if let Some(conditional) = self.ctx.arena.get_conditional_type(parent_node)
-                && self.type_node_contains_infer_binding_named(conditional.extends_type, name)
-            {
-                return true;
-            }
-            current = parent;
-        }
-        false
-    }
-
-    fn type_node_contains_infer_binding_named(&self, node_idx: NodeIndex, name: &str) -> bool {
-        let Some(node) = self.ctx.arena.get(node_idx) else {
-            return false;
-        };
-        if node.kind == syntax_kind_ext::TEMPLATE_LITERAL_TYPE_SPAN {
-            return self.ctx.arena.get_template_span(node).is_some_and(|span| {
-                self.type_node_contains_infer_binding_named(span.expression, name)
-            });
-        }
-        if node.kind == syntax_kind_ext::INFER_TYPE {
-            let Some(infer_data) = self.ctx.arena.get_infer_type(node) else {
-                return false;
-            };
-            if let Some(type_param_node) = self.ctx.arena.get(infer_data.type_parameter)
-                && let Some(type_param) = self.ctx.arena.get_type_parameter(type_param_node)
-                && let Some(name_node) = self.ctx.arena.get(type_param.name)
-                && let Some(identifier) = self.ctx.arena.get_identifier(name_node)
-                && identifier.escaped_text == name
-            {
-                return true;
-            }
-            return self.type_node_contains_infer_binding_named(infer_data.type_parameter, name);
-        }
-        if node.kind == syntax_kind_ext::TYPE_PARAMETER
-            && let Some(type_param) = self.ctx.arena.get_type_parameter(node)
-        {
-            return (type_param.constraint != NodeIndex::NONE
-                && self.type_node_contains_infer_binding_named(type_param.constraint, name))
-                || (type_param.default != NodeIndex::NONE
-                    && self.type_node_contains_infer_binding_named(type_param.default, name));
-        }
-
-        self.ctx
-            .arena
-            .get_children(node_idx)
-            .into_iter()
-            .any(|child_idx| self.type_node_contains_infer_binding_named(child_idx, name))
     }
 
     /// Walk `extends_type` collecting every `infer X` binding and push each as a
